add user permission

+++ chat/service/agent.py
@@ -1,21 +1,29 @@
+from aiohttp.test_utils import REUSE_ADDRESS
+
+from database import AgentTable
 from database.dao.agent import AgentDao
+from database.models.user import AdminUser, SystemUser
 from loguru import logger
+
+from test.test_tool import message
+from type.schemas import resp_200, resp_500
 
 
 class AgentService:
 
     @classmethod
-    def create_agent(cls, name: str, description: str, logo: str, parameter: str, type: str = "openai", code: str = "",
-                     isCustom: bool = True):
+    def create_agent(cls, name: str, description: str, logo: str, user_id: int,
+                     parameter: str, type: str = "openai", code: str = "", is_custom: bool = True):
         try:
-            agentId = AgentDao.create_agent(name=name,
+            agent_id = AgentDao.create_agent(name=name,
                                             description=description,
                                             logo=logo,
                                             parameter=parameter,
                                             type=type,
                                             code=code,
-                                            isCustom=isCustom)
-            return agentId
+                                            user_id=user_id,
+                                            is_custom=is_custom)
+            return agent_id
         except Exception as err:
             logger.error(f"create agent is appear error: {err}")
 
@@ -42,9 +50,9 @@ class AgentService:
             logger.error(f"select agent by type is appear error: {err}")
 
     @classmethod
-    def select_agent_by_custom(cls, isCustom):
+    def select_agent_by_custom(cls, is_custom):
         try:
-            data = AgentDao.select_agent_by_custom(isCustom=isCustom)
+            data = AgentDao.select_agent_by_custom(is_custom=is_custom)
             result = []
             for item in data:
                 result.append(item[0])
@@ -91,28 +99,47 @@ class AgentService:
             logger.error(f"get agent by name and type appear error: {err}")
 
     @classmethod
-    def update_agent_by_id(cls, id: str, name: str, description: str, logo: str, parameter: str, code: str):
+    def update_agent_by_id(cls, id: str, name: str, description: str, user_id: int, logo: str, parameter: str, code: str, type: str=None):
         try:
-            AgentDao.update_agent_by_id(id=id,
-                                        name=name,
-                                        logo=logo,
-                                        description=description,
-                                        parameter=parameter,
-                                        code=code)
+            # 需要判断是否有权限，管理员随意
+            if user_id == AdminUser or user_id == cls.get_agent_user_id(agent_id=id):
+                AgentDao.update_agent_by_id(id=id,
+                                            name=name,
+                                            logo=logo,
+                                            type=type,
+                                            description=description,
+                                            parameter=parameter,
+                                            code=code)
+                return resp_200(message='update agent success')
+            else:
+                return resp_500(message='no permission exec')
         except Exception as err:
             logger.error(f"update agent by id appear error: {err}")
 
     @classmethod
-    def delete_agent_by_id(cls, id: str):
+    def get_agent_user_id(cls, agent_id: str):
         try:
-            AgentDao.delete_agent_by_id(id=id)
+            agent = AgentDao.get_agent_user_id(agent_id=agent_id)
+            return agent.user_id
+        except Exception as err:
+            logger.error(f'get agent user id Err: {err}')
+
+    @classmethod
+    def delete_agent_by_id(cls, id: str, user_id: int):
+        try:
+            # 需要判断是否有权限，管理员随意
+            if user_id == AdminUser or user_id == cls.get_agent_user_id(agent_id=id):
+                AgentDao.delete_agent_by_id(id=id)
+                return resp_200(message='delete success')
+            else:
+                return resp_500(message='no permission exec')
         except Exception as err:
             logger.error(f"delete agent by id appear: {err}")
 
     @classmethod
-    def search_agent_name(cls, name: str):
+    def search_agent_name(cls, name: str, user_id: int):
         try:
-            data = AgentDao.search_agent_name(name=name)
+            data = AgentDao.search_agent_name(name=name, user_id=user_id)
             result = []
             for item in data:
                 result.append(item[0])
@@ -135,6 +162,29 @@ class AgentService:
     def check_name_iscustom(cls, name: str):
         try:
             data = AgentDao.select_agent_by_name(name)
-            return data[0][0].isCustom
+            return data[0][0].is_custom
         except Exception as err:
             logger.error(f"get code by name is appear error: {err}")
+
+    @classmethod
+    def get_personal_agent_by_user_id(cls, user_id: int):
+        try:
+            data = AgentDao.get_agent_by_user_id(user_id=user_id)
+            result = []
+            for item in data:
+                result.append(item[0])
+            return result
+        except Exception as err:
+            logger.error(f'get personal agent by user id Err: {err}')
+
+    @classmethod
+    def get_all_agent_by_user_id(cls, user_id: int):
+        try:
+            system_data = AgentDao.get_agent_by_user_id(user_id=SystemUser)
+            user_data = AgentDao.get_agent_by_user_id(user_id=user_id)
+            result = []
+            for item in set(system_data + user_data):
+                result.append(item[0])
+            return result
+        except Exception as err:
+            logger.error(f'get all agent by user id Err: {err}')
